+++ ricaun.Revit.UI.Tasks/IRevitTask.cs
@@ -1,5 +1,6 @@
 ﻿using Autodesk.Revit.UI;
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace ricaun.Revit.UI.Tasks
@@ -12,6 +13,10 @@ namespace ricaun.Revit.UI.Tasks
         /// <summary>
         /// Run code in Revit context.
         /// </summary>
-        public Task<TResult> Run<TResult>(Func<UIApplication, TResult> function);
+        /// <typeparam name="TResult"></typeparam>
+        /// <param name="function"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        public Task<TResult> Run<TResult>(Func<UIApplication, TResult> function, CancellationToken cancellationToken);
     }
 }