+++ ricaun.Revit.UI.Tasks/RevitTask.cs
@@ -1,5 +1,6 @@
 ﻿using Autodesk.Revit.UI;
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace ricaun.Revit.UI.Tasks
@@ -23,9 +24,9 @@ namespace ricaun.Revit.UI.Tasks
         /// <summary>
         /// Run code in Revit context.
         /// </summary>
-        public Task<TResult> Run<TResult>(Func<UIApplication, TResult> function)
+        public Task<TResult> Run<TResult>(Func<UIApplication, TResult> function, CancellationToken cancellationToken)
         {
-            return revitTaskService.Run(function);
+            return revitTaskService.Run(function, cancellationToken);
         }
     }