+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] / 2024-01-27
-- First Release
+### Features
+- Support `RevitTaskService` with `IRevitTask`
+- Support `CancellationToken` in `IRevitTask`.
+### Added
 - Add `RevitTaskService`
 - Add `IRevitTask`
 - Add `IRevitTaskExtension`
+
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0